pyproject.toml
[project]
name = "telegram-ai-management-bot"
version = "0.1.0"
description = "Telegram management bot with AI task planning via Groq DeepSeek R1"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "aiogram>=3.21.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1"
]
